feature-이동석
드롭다운.jsx 스펠링 오타 수정
css 코드 수정
